Dorada za kontrolu lagera na inventaru

+++ src/Purchase/UseCase/DeletePurchaseUseCase.php
@@ -2,7 +2,9 @@
 
 namespace App\Purchase\UseCase;
 
+use App\Consumption\Exception\InvalidConsumptionException;
 use App\Core\Service\ContextService;
+use App\Inventory\Repository\InventoryRepository;
 use App\Purchase\Entity\Purchase;
 use App\Purchase\Repository\PurchaseRepository;
 use Exception;
@@ -16,15 +18,18 @@ class DeletePurchaseUseCase
 {
     private ContextService $contextService;
     private PurchaseRepository $purchaseRepository;
+    private InventoryRepository $inventoryRepository;
 
-    public function __construct(ContextService $contextService, PurchaseRepository $purchaseRepository)
+    public function __construct(ContextService $contextService, PurchaseRepository $purchaseRepository, InventoryRepository $inventoryRepository)
     {
         $this->contextService = $contextService;
         $this->purchaseRepository = $purchaseRepository;
+        $this->inventoryRepository = $inventoryRepository;
     }
 
     public function execute(string $purchaseId): void
     {
+        /** @var Purchase $purchase */
         $purchase = $this->purchaseRepository->findOneBy(["id" => $purchaseId]);
 
         if (is_null($purchase)) {
@@ -35,7 +40,11 @@ class DeletePurchaseUseCase
             throw new AccessDeniedHttpException();
         }
 
-        // TODO: provjera je li ima dovoljno na lageru
+
+        $currentStock = $this->inventoryRepository->findStockForSupply($purchase->getFarm(), $purchase->getSupply());
+        if ($purchase->getAmount() > $currentStock) {
+            throw new InvalidConsumptionException("Neće biti dovoljno na lageru!");
+        }
 
         try {
             $this->purchaseRepository->delete($purchase, true);
